move attributes to their own place
