+++ .gitignore
@@ -58,3 +58,4 @@ typings/
 .env
 
 domain
+!src/domain
