+++ src/Domain/Objects/AttributeParser/Attributes/ArrayAttribute.js
