+++ src/Domain/Objects/AttributeParser/Attributes/BooleanAttribute.js
