+++ src/Domain/Objects/AttributeParser/Attributes/FloatAttribute.js
