+++ src/Domain/Objects/AttributeParser/Attributes/IntegerAttribute.js
