+++ src/Domain/Objects/AttributeParser/Attributes/JSONAttribute.js
