+++ src/Domain/Objects/AttributeParser/Attributes/ObjectAttribute.js
