+++ src/Domain/Objects/AttributeParser/Attributes/StringAttribute.js
